requirements.txt
markdown==2.6.7
django==1.10.3
selenium==3.0.1
PyYAML==3.12
gunicorn==19.6.0
whitenoise==3.2.2
pyPodcastParser==2.0.0
requests==2.18.1
